# Conflicts:
#	src/main/java/com/ii/testautomation/controllers/SchedulingController.java
#	src/main/java/com/ii/testautomation/dto/request/SchedulingRequest.java
#	src/main/java/com/ii/testautomation/dto/response/SchedulingResponse.java
#	src/main/java/com/ii/testautomation/entities/Scheduling.java
#	src/main/java/com/ii/testautomation/repositories/SchedulingRepository.java
#	src/main/java/com/ii/testautomation/service/SchedulingService.java
#	src/main/java/com/ii/testautomation/service/impl/SchedulingServiceImpl.java
#	src/main/java/com/ii/testautomation/utils/Constants.java
#	src/main/java/com/ii/testautomation/utils/EndpointURI.java
#	src/main/java/com/ii/testautomation/utils/StatusCodeBundle.java
#	src/main/resources/MessagesAndCodes.properties

+++ src/main/java/com/ii/testautomation/service/impl/ExecutionHistoryServiceImpl.java
@@ -2,6 +2,7 @@ package com.ii.testautomation.service.impl;
 
 import com.ii.testautomation.dto.response.ExecutionHistoryResponse;
 import com.ii.testautomation.entities.ExecutionHistory;
+import com.ii.testautomation.entities.Project;
 import com.ii.testautomation.repositories.ExecutionHistoryRepository;
 import com.ii.testautomation.repositories.ProjectRepository;
 import com.ii.testautomation.repositories.TestGroupingRepository;
@@ -13,16 +14,9 @@ import org.springframework.stereotype.Service;
 
 import java.io.File;
 import java.io.IOException;
-import java.io.File;
-import java.io.IOException;
-import javax.swing.text.html.HTML;
-import javax.swing.text.html.HTMLDocument;
-import java.io.*;
-import java.net.URL;
-import java.nio.charset.StandardCharsets;
 import java.nio.file.Files;
 import java.nio.file.Path;
-import java.sql.Date;
+import java.sql.Timestamp;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -78,27 +72,42 @@ public class ExecutionHistoryServiceImpl implements ExecutionHistoryService {
     }
 
     @Override
-    public void deleteExecutionHistory(Long id, Long projectId) {
-        executionHistoryRepository.deleteById(id);
-        String historyReport = fileFolder + File.separator + projectRepository.findById(id).get().getName() + File.separator + executionHistoryRepository.findById(id).get().getReportName().toString() + ".html";
-        deleteReport(historyReport);
-    }
-
-    @Override
-    public List<ExecutionHistoryResponse> getByTestGroupingIdWithDate(Long id,Date date) {
-        List<ExecutionHistoryResponse> executionHistoryResponseList = new ArrayList<>();
-        List<ExecutionHistory> executionHistoryList = new ArrayList<>();
-        return executionHistoryResponseList;
+    public boolean deleteExecutionHistory(Long id, Long projectId) {
+        Project projectOptional = projectRepository.findById(projectId).get();
+        if (projectOptional != null) {
+            String historyReport = projectOptional.getProjectPath() + File.separator + executionHistoryRepository.findById(id).get().getReportName().toString() + ".html";
+            if (historyReport != null && !historyReport.isEmpty()) {
+                if (deleteReport(historyReport)) {
+                    executionHistoryRepository.deleteById(id);
+                    return true;
+                }
+            }
+        }
+        return false;
     }
 
     private boolean deleteReport(String filePath) {
         File fileToDelete = new File(filePath);
-
         if (fileToDelete.exists() && fileToDelete.isFile()) {
             return fileToDelete.delete();
         }
         return false;
     }
+
+    @Override
+    public List<ExecutionHistoryResponse> executionHistoryDateFilter(Long id, Timestamp startDate, Timestamp endDate) {
+        List<ExecutionHistoryResponse> executionHistoryResponseList = new ArrayList<>();
+        List<ExecutionHistory> executionHistoryList = executionHistoryRepository.findByTestGroupingIdAndCreatedAtBetween(id, startDate, endDate);
+        for (ExecutionHistory executionHistory : executionHistoryList) {
+            ExecutionHistoryResponse executionHistoryResponse = new ExecutionHistoryResponse();
+            BeanUtils.copyProperties(executionHistory, executionHistoryResponse);
+            executionHistoryResponse.setTestGroupingId(id);
+            executionHistoryResponseList.add(executionHistoryResponse);
+        }
+
+        return executionHistoryResponseList;
+    }
+
 
     @Override
     public boolean existByTestGropingId(Long id) {
